requirements.txt
python-magic~=0.4
Pillow~=9.5